+++ extensions/amp-a4a/0.1/friendly-frame-util.js
@@ -17,10 +17,8 @@
 import {A4AVariableSource} from '../../amp-a4a/0.1/a4a-variable-source';
 import {createElementWithAttributes} from '../../../src/dom';
 import {dict} from '../../../src/utils/object';
-import {
-  getExtensionsFromMetadata,
-  installFriendlyIframeEmbed,
-} from '../../../src/friendly-iframe-embed';
+import {getExtensionsFromMetadata} from './amp-ad-utils';
+import {installFriendlyIframeEmbed} from '../../../src/friendly-iframe-embed';
 import {installUrlReplacementsForEmbed} from '../../../src/service/url-replacements-impl';
 import {setStyle} from '../../../src/style';
