+++ test/unit/test-friendly-iframe-embed.js
@@ -22,8 +22,6 @@ import {FakeWindow} from '../../testing/fake-dom';
 import {
   FriendlyIframeEmbed,
   Installers,
-  extensionsHasElement,
-  getExtensionsFromMetadata,
   installFriendlyIframeEmbed,
   mergeHtmlForTesting,
   preloadFriendlyIframeEmbedExtensionIdsDeprecated,
@@ -1542,64 +1540,5 @@ describes.realWin('installExtensionsInEmbed', {amp: true}, (env) => {
     // Extension elements are stubbed immediately, but registered only
     // after extension is loaded.
     expect(iframeWin.__AMP_EXTENDED_ELEMENTS['amp-test']).to.equal(AmpTest);
-  });
-});
-
-describe('extensionsHasElement', () => {
-  it('should return true if containing extension', () => {
-    const extensions = [
-      {
-        'custom-element': 'amp-cats',
-        src: 'https://cdn.ampproject.org/v0/amp-cats-0.1.js',
-      },
-    ];
-    expect(extensionsHasElement(extensions, 'amp-cats')).to.be.true;
-  });
-
-  it('should return false if it does not contain extension', () => {
-    const extensions = [
-      {
-        'custom-element': 'amp-cats',
-        src: 'https://cdn.ampproject.org/v0/amp-cats-0.1.js',
-      },
-    ];
-    expect(extensionsHasElement(extensions, 'amp-dogs')).to.be.false;
-  });
-
-  it('should return false if empty array', () => {
-    const extensions = [];
-    expect(extensionsHasElement(extensions, 'amp-dogs')).to.be.false;
-  });
-});
-
-describe('getExtensionsFromMetadata', () => {
-  it('should return extension name and version', () => {
-    const metadata = {
-      extensions: [
-        {
-          'custom-element': 'amp-analytics',
-          'src': 'https://cdn.ampproject.org/v0/amp-analytics-0.1.js',
-        },
-        {
-          'custom-element': 'amp-mustache',
-          'src': 'https://cdn.ampproject.org/v0/amp-mustache-1.0.js',
-        },
-      ],
-    };
-    const extensions = getExtensionsFromMetadata(metadata);
-    expect(extensions).to.deep.include({
-      extensionId: 'amp-analytics',
-      extensionVersion: '0.1',
-    });
-    expect(extensions).to.deep.include({
-      extensionId: 'amp-mustache',
-      extensionVersion: '1.0',
-    });
-  });
-
-  it('should handle no `extensions` key in metadata', () => {
-    const metadata = {};
-    const extensions = getExtensionsFromMetadata(metadata);
-    expect(extensions).to.eql([]);
   });
 });
